OneDrive/Desktop/03_Image_Classification_CNN/image_classification_cnn/requirements.txt
tensorflow>=2.13.0
streamlit>=1.32.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
